requirements.txt
adbc_driver_manager==1.9.0
adbc_driver_postgresql==1.9.0
adbc_driver_sqlite==1.9.0
anywidget==0.9.21
AppKit==0.2.8
atheris==2.3.0
authlib==1.6.5
beautifulsoup4==4.14.2
bokeh==3.8.1
boto3==1.40.74
botocore==1.40.74
Brotli==1.1.0
brotlicffi==1.1.0.0
chardet==5.2.0
ConfigParser==7.2.0
contextlib2==21.6.0
cryptography==46.0.3
cudf==0.6.1.post1
cupy==13.6.0
cycler==0.12.1
Cython==3.2.1
dask==2025.11.0
dask_expr==2.0.0
defusedxml==0.7.1
dl==0.1.0
docutils==0.22.3
duckdb==1.4.2
fastparquet==2024.11.0
filelock==3.20.0
Foundation==0.1.0a0.dev1
fqdn==1.5.1
fsspec==2025.10.0
gitdb_speedups==0.1.0
graphviz==0.21
h2==4.3.0
HTMLParser==0.0.2
huggingface_hub==0.35.3
hypothesis==6.148.1
ibis==3.3.0
ipython==8.12.3
ipywidgets==8.1.7
isoduration==20.11.0
jnius==1.1.0
JPype1==1.6.0
jsonpath_ng==1.7.0
jsonpointer==3.0.0
keyring==25.7.0
lxml==5.4.0
matplotlib==3.7.1
mock==5.2.0
modin==0.37.1
mtrand==0.1
mypy==1.18.2
numarray==1.5.1
Numeric==24.2
numexpr==2.14.1
odfpy==1.4.1
olefile==0.47
opencv_python==4.11.0.86
openpyxl==3.1.5
plotly==6.4.0
polars==1.35.2
psutil==7.0.0
py==1.11.0
pycares==4.11.0
pycurl==7.45.7
PyInstaller==6.16.0
pyodide==0.0.2
pyOpenSSL==25.3.0
pyperf==2.9.0
PyQt4==4.11.4
PyQt5==5.15.11
PyQt6==6.10.0
PySide6==6.10.0
pyspark==4.0.1
pytest==9.0.1
python_calamine==0.5.4
pywin32==311
pyxlsb==1.0.10
PyYAML==6.0.2
PyYAML==6.0.3
qtpy==2.4.3
railroad==0.5.0
redis==7.0.1
rfc3339_validator==0.1.4
rfc3986_validator==0.1.1
rfc3987==1.3.8
rfc3987_syntax==1.1.0
s3fs==2025.10.0
scikit_learn==1.2.2
scipy==1.16.3
scipy_doctest==2.0.1
seaborn==0.13.2
sets==0.3.2
setuptools_scm==9.2.2
simplejson==3.20.2
sniffio==1.3.1
snowflake==1.9.0
Sphinx==8.2.3
SQLAlchemy==2.0.44
sqlframe==3.43.8
streamlit_pdf==2.0.0
sympy==1.14.0
tables==3.10.2
thread==2.0.5
threadpoolctl==3.6.0
traitlets==5.14.3
trio==0.32.0
trove_classifiers==2025.11.14.15
uri_template==1.3.0
urllib3_secure_extra==0.1.0
vegafusion==2.0.3
webcolors==25.10.0
xarray==2025.10.1
xlrd==2.0.2
xlsxwriter==3.2.9
xmlrpclib==1.0.1
zstandard==0.25.0
